+++ xtls/tls.go
@@ -1,4 +1,4 @@
-package tls
+package xtls
 
 import (
 	"bytes"
